ex01 finished

+++ cpp06/ex01/Serializer.hpp
@@ -4,8 +4,7 @@
 
 #include <iostream>
 #include <string>
-#include "Data.h"
-
+#include "Data.hpp"
 
 class Serializer
 {
@@ -16,8 +15,8 @@ class Serializer
 		~Serializer();
 		Serializer & operator=(Serializer const & src);
 
-		static uintptr_t serialize(Data* ptr);
-		static Data* deserialize(uintptr_t raw);
+		static unsigned long serialize(Data* ptr);
+		static Data* deserialize(unsigned long raw);
 };
 
 #endif
